Fix stylying and logic of plugin
