+++ superset-frontend/src/Superstructure/chart/chartAction.js
@@ -108,9 +108,6 @@ const legacyChartDataRequest = async (
   const allowDomainSharding =
     // eslint-disable-next-line camelcase
     domainShardingEnabled && requestParams?.dashboard_id;
-  console.log('legacyChartDataRequest PLUGIN')
-  console.log('endpointType', endpointType)
-
   const url = getExploreUrl({
     formData,
     endpointType,
@@ -132,7 +129,6 @@ const legacyChartDataRequest = async (
   bodyFormData.append('form_data', JSON.stringify(newFormData));
 
   if (isFeatureEnabled(FeatureFlag.CLIENT_CACHE)) {
-    console.log('PLUGIN isFeatureEnabled true', url)
     return API_HANDLER.SupersetClientNoApi({
       method: 'get',
       url,
@@ -141,7 +137,6 @@ const legacyChartDataRequest = async (
       headers: { 'Content-Type': 'multipart/form-data' },
     });
   }
-  console.log('PLUGIN isFeatureEnabled false', url)
   return API_HANDLER.SupersetClientNoApi({
     method: 'post',
     url,
@@ -180,6 +175,8 @@ const v1ChartDataRequest = async (
     ownState,
   });
 
+  console.log('chart action PLUGIN')
+
   // The dashboard id is added to query params for tracking purposes
   const { slice_id: sliceId } = formData;
   const { dashboard_id: dashboardId } = requestParams;
@@ -198,8 +195,6 @@ const v1ChartDataRequest = async (
     qs,
     allowDomainSharding,
   }).toString();
-
-  console.log('buildV1ChartDataPayload', url);
 
   return API_HANDLER.SupersetClient({ method: 'post', url, body: payload });
 };
@@ -225,8 +220,6 @@ export async function getChartDataRequest({
       credentials: 'include',
     };
   }
-
-  console.log('getChartDataRequest PLUGIN')
 
   if (shouldUseLegacyApi(formData)) {
     return legacyChartDataRequest(
