+++ superset-frontend/src/Superstructure/components/MainRoute/index.tsx
@@ -1,9 +1,7 @@
 import { hot } from 'react-hot-loader/root';
 import React from 'react';
 
-import { AnalyticsMainWrapper, StyledH1, Alert } from './styles';
-
-import { MAIN_TITLE } from './constants';
+import { AnalyticsMainWrapper, Alert } from './styles';
 
 import {
   InfoIcon,
@@ -24,7 +22,6 @@ const AnalyticsMain = () => {
   return (
     <AnalyticsMainWrapper>
       <RowWrapper>
-        <StyledH1>{MAIN_TITLE}</StyledH1>
       </RowWrapper>
       <RowWrapper>
         <ColumnWrapper classes="col-sm-12 col-md-6">
