+++ superset-frontend/src/Superstructure/components/index.ts
@@ -1,5 +1,5 @@
 import { ErrorBanner } from './ErrorBanner';
 import { Version } from './Version';
-import Loading from './Loading';
+import Loading from 'src/components/Loading';
 
 export { ErrorBanner, Version, Loading };
