+++ superset-frontend/src/Superstructure/dashboard/containers/Chart.jsx
@@ -20,7 +20,7 @@ import getFormDataWithExtraFilters from 'src/dashboard/util/charts/getFormDataWi
 import Chart from 'src/Superstructure/dashboard/components/gridComponents/Chart';
 import { PLACEHOLDER_DATASOURCE } from 'src/dashboard/constants';
 
-console.log('ChartXX grid components is used here')
+console.log('Chart container PLUGIN')
 
 const EMPTY_FILTERS = {};
