+++ superset-frontend/src/Superstructure/dashboard/containers/DashboardPage.tsx
@@ -1,7 +1,7 @@
 // DODO-changed
 import React, { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import Loading from 'src/Superstructure/components/Loading';
+import Loading from 'src/components/Loading';
 import {
   fetchDashboard,
   fetchDashboardCharts,
