+++ superset-frontend/src/Superstructure/stylesheets/less/index.less
@@ -31,5 +31,5 @@ header {
 
 .active-link {
   color: #fff !important;
-  background: #555 !important;
+  background: #6c757d !important;
 }
