+++ superset-frontend/src/chart/chartAction.js
@@ -174,6 +174,8 @@ const v1ChartDataRequest = async (
     ownState,
   });
 
+  console.log('chart action STANDALONE')
+
   // The dashboard id is added to query params for tracking purposes
   const { slice_id: sliceId } = formData;
   const { dashboard_id: dashboardId } = requestParams;
