+++ superset-frontend/src/components/Loading/index.tsx
@@ -1,21 +1,4 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
+// DODO-changed
 
 import React from 'react';
 import { styled } from '@superset-ui/core';
@@ -32,7 +15,7 @@ export interface Props {
   image?: string;
 }
 
-const LoaderImg = styled.img`
+const LoaderDiv = styled.div`
   z-index: 99;
   width: 50px;
   position: relative;
@@ -55,19 +38,73 @@ const LoaderImg = styled.img`
     transform: translate(-50%, -50%);
   }
 `;
-export default function Loading({
-  position = 'floating',
-  image = '/static/assets/images/loading.gif',
-  className,
-}: Props) {
+
+const LdsRing = styled.div`
+  display: flex;
+  position: relative;
+  width: 40px;
+  height: 40px;
+
+  div {
+    position: absolute;
+    top: 15px;
+    width: 8px;
+    height: 8px;
+    border-radius: 50%;
+    background: #ff6900;
+    animation-timing-function: cubic-bezier(0, 1, 1, 0);
+  }
+  div:nth-child(1) {
+    left: 4px;
+    animation: lds-ellipsis1 0.6s infinite;
+  }
+  div:nth-child(2) {
+    left: 4px;
+    animation: lds-ellipsis2 0.6s infinite;
+  }
+  div:nth-child(3) {
+    left: 16px;
+    animation: lds-ellipsis2 0.6s infinite;
+  }
+  div:nth-child(4) {
+    left: 28px;
+    animation: lds-ellipsis3 0.6s infinite;
+  }
+  @keyframes lds-ellipsis1 {
+    0% {
+      transform: scale(0);
+    }
+    100% {
+      transform: scale(1);
+    }
+  }
+  @keyframes lds-ellipsis3 {
+    0% {
+      transform: scale(1);
+    }
+    100% {
+      transform: scale(0);
+    }
+  }
+  @keyframes lds-ellipsis2 {
+    0% {
+      transform: translate(0, 0);
+    }
+    100% {
+      transform: translate(12px, 0);
+    }
+  }
+`;
+
+export default function Loading({ position = 'floating', className }: Props) {
   return (
-    <LoaderImg
-      className={cls('loading', position, className)}
-      alt="Loading..."
-      src={image}
-      role="status"
-      aria-live="polite"
-      aria-label="Loading"
-    />
+    <LoaderDiv className={cls('loading', position, className)}>
+      <LdsRing>
+        <div />
+        <div />
+        <div />
+        <div />
+      </LdsRing>
+    </LoaderDiv>
   );
 }
