+++ superset-frontend/src/dashboard/components/SliceHeaderControls/index.tsx
@@ -280,7 +280,7 @@ class SliceHeaderControls extends React.PureComponent<
           </Menu.Item>
         )}
 
-        {this.props.supersetCanExplore && (
+        {(this.props.supersetCanExplore && !process.env.business) && (
           <Menu.Item key={MENU_KEYS.EXPLORE_CHART}>
             <a href={this.props.exploreUrl} rel="noopener noreferrer">
               {t('View chart in Explore')}
@@ -288,7 +288,7 @@ class SliceHeaderControls extends React.PureComponent<
           </Menu.Item>
         )}
 
-        {this.props.supersetCanExplore && (
+        {(this.props.supersetCanExplore && !process.env.business) && (
           <Menu.Item key={MENU_KEYS.VIEW_QUERY}>
             <ModalTrigger
               triggerNode={
@@ -303,7 +303,7 @@ class SliceHeaderControls extends React.PureComponent<
           </Menu.Item>
         )}
 
-        {supersetCanShare && (
+        {(supersetCanShare && !process.env.business) && (
           <ShareMenuItems
             url={getDashboardUrl({
               pathname: window.location.pathname,
