+++ superset-frontend/src/dashboard/components/gridComponents/index.js
@@ -9,9 +9,8 @@ import {
   TAB_TYPE,
   TABS_TYPE,
 } from '../../util/componentTypes';
-// TODO: fix dodo
+
 import ChartHolder from 'src/dashboard/components/gridComponents/ChartHolder';
-// import ChartHolder from 'src/Superstructure/dashboard/components/gridComponents/ChartHolder';
 
 import Markdown from './Markdown';
 import Column from './Column';
@@ -21,9 +20,6 @@ import Row from './Row';
 import Tab from './Tab';
 import TabsConnected from './Tabs';
 
-console.log('gridComponents process.env.business', process.env.business);
-// TODO: fix dodo
-// export { default as ChartHolder } from 'src/Superstructure/dashboard/components/gridComponents/ChartHolder';
 export { default as ChartHolder } from 'src/dashboard/components/gridComponents/ChartHolder';
 
 export { default as Markdown } from './Markdown';
@@ -35,7 +31,6 @@ export { default as Tab } from './Tab';
 export { default as Tabs } from './Tabs';
 
 export const componentLookup = {
-  // [CHART_TYPE]: !process.env.business ? ChartHolder : ChartHolderPlugin,
   [CHART_TYPE]: ChartHolder,
   [MARKDOWN_TYPE]: Markdown,
   [COLUMN_TYPE]: Column,
