+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterControls/FilterValue.tsx
@@ -1,21 +1,4 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
+// DODO-changed
 import React, { useEffect, useRef, useState } from 'react';
 import {
   QueryFormData,
@@ -31,6 +14,7 @@ import {
 import { useDispatch, useSelector } from 'react-redux';
 import { areObjectsEqual } from 'src/reduxUtils';
 import { getChartDataRequest } from 'src/chart/chartAction';
+import { getChartDataRequest as getChartDataRequestPlugin } from 'src/Superstructure/chart/chartAction';
 import Loading from 'src/components/Loading';
 import BasicErrorAlert from 'src/components/ErrorMessage/BasicErrorAlert';
 import { FeatureFlag, isFeatureEnabled } from 'src/featureFlags';
@@ -117,7 +101,8 @@ const FilterValue: React.FC<FilterProps> = ({
         return;
       }
       setIsRefreshing(true);
-      getChartDataRequest({
+
+      !process.env.business ? getChartDataRequest : getChartDataRequestPlugin({
         formData: newFormData,
         force: false,
         requestParams: { dashboardId: 0 },
